python-api/requirements.txt
flask
tensorflow
numpy
pillow
flask-cors
gdown